+++ Source/App/AppController+URL.swift
@@ -113,7 +113,9 @@ extension AppController {
     // reset the root controller, not push into a child feature controller
     func pushViewController(for contentType: ContentType, with identity: Identity) {
         guard contentType == .about else { return }
-        let view = IdentityView(identity: identity).environmentObject(BotRepository.shared)
+        let view = IdentityView(identity: identity)
+            .environmentObject(BotRepository.shared)
+            .environmentObject(AppController.shared)
         let controller = UIHostingController(rootView: view)
         self.push(controller, animated: true)
     }
@@ -135,7 +137,9 @@ extension AppController {
     }
 
     func pushChannelViewController(for hashtag: String) {
-        let view = HashtagView(hashtag: Hashtag.named(hashtag)).environmentObject(BotRepository.shared)
+        let view = HashtagView(hashtag: Hashtag.named(hashtag))
+            .environmentObject(BotRepository.shared)
+            .environmentObject(AppController.shared)
         self.push(UIHostingController(rootView: view))
     }