+++ Source/Controller/AboutTableViewDataSource.swift
@@ -95,7 +95,9 @@ extension AboutTableViewDataSource: UITableViewDelegate {
         Analytics.shared.trackDidSelectItem(kindName: "identity")
         let identity = self.identities[indexPath.row]
         let controller = UIHostingController(
-            rootView: IdentityView(identity: identity).environmentObject(BotRepository.shared)
+            rootView: IdentityView(identity: identity)
+                .environmentObject(BotRepository.shared)
+                .environmentObject(AppController.shared)
         )
         let targetController = self.delegate?.navigationController
         targetController?.pushViewController(controller, animated: true)