+++ Source/UI/About/MessageListView.swift
@@ -12,14 +12,17 @@ import SwiftUI
 
 struct MessageListView: View {
 
-    @State
-    var messages = [Message]()
-
     var strategy: FeedStrategy
 
     @EnvironmentObject
     private var botRepository: BotRepository
 
+    @EnvironmentObject
+    private var appController: AppController
+
+    @State
+    private var messages = [Message]()
+    
     @State
     private var isLoading = false
 
@@ -38,9 +41,9 @@ struct MessageListView: View {
                     ForEach(messages, id: \.self) { message in
                         Button {
                             if let contact = message.content.contact {
-                                AppController.shared.open(identity: contact.contact)
+                                appController.open(identity: contact.contact)
                             } else {
-                                AppController.shared.open(identifier: message.id)
+                                appController.open(identifier: message.id)
                             }
                         } label: {
                             MessageView(message: message)
@@ -152,15 +155,15 @@ struct MessageListView_Previews: PreviewProvider {
     static var previews: some View {
         Group {
             VStack {
-                MessageListView(messages: [], strategy: NoHopFeedAlgorithm(identity: .null))
+                MessageListView(strategy: StaticAlgorithm(messages: []))
                 ScrollView {
-                    MessageListView(messages: [sample, another], strategy: NoHopFeedAlgorithm(identity: .null))
+                    MessageListView(strategy: StaticAlgorithm(messages: [sample, another]))
                 }
             }
             VStack {
-                MessageListView(messages: [], strategy: NoHopFeedAlgorithm(identity: .null))
+                MessageListView(strategy: StaticAlgorithm(messages: []))
                 ScrollView {
-                    MessageListView(messages: [sample, another], strategy: NoHopFeedAlgorithm(identity: .null))
+                    MessageListView(strategy: StaticAlgorithm(messages: [sample, another]))
                 }
             }
             .preferredColorScheme(.dark)
@@ -168,5 +171,6 @@ struct MessageListView_Previews: PreviewProvider {
         .padding()
         .background(Color.cardBackground)
         .environmentObject(BotRepository.fake)
+        .environmentObject(AppController.shared)
     }
 }