+++ CHANGELOG.md
@@ -6,7 +6,11 @@ We define "Noteworthy changes" as 1) user-facing features or bugfixes 2) signifi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.12] 2022-12-09
+## [1.3.12] 2022-12-13
+
+- Fixed a bug that caused hashtags to not display any posts. #1021
+
+## [1.3.11] 2022-12-09
 
 - Added new redesigned Home screen. #974
 - Added new redesigned Hashtag screen. #975
@@ -17,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Disabled EBT when the user is restoring their feed to work around a bug in our pubs
 - Added new redesigned Profile screen. #681 #976 
 - Fixed a bug in macOS that prevented users from load the identity from the keychain. #978
+- Publish announce messages for room aliases #899
 
 ## [1.3.9] 2022-11-10